General fixes (incl. race condition), code cleanup
This commit contains general fixes for the relay and adds no new
functionality.  In particular, this commit protects against a potential
race condition that may arise due to the outfile next read position
being based on the outfile's size instead of the actual last read
position; this meant that there was a very small chance that the outfile
would be updated in the short time between the readLine() loop ending
and the filesize being taken, and those updates would not been seen by
the relay.

Change-Id: I1fe0942ea9fa1485ae6cb5acd434d70d897fa8c5